build(nightmare): add e2e config and tasks
