+++ src/client/app/framework/testing/e2e.ts
@@ -1,0 +1,4 @@
+// module
+import { ngE2e } from './shorthand/ng-e2e';
+
+export const e2e = ngE2e;
