+++ src/client/app/framework/testing/shorthand/ng-e2e.ts
@@ -1,0 +1,20 @@
+// module
+import { TestApi } from './ng-base';
+
+export const ngE2e: TestApi = {
+  ae: afterEach, // shorthand afterEach
+  afterEach,
+  describe,
+  fdescribe,
+  xdescribe,
+  be: beforeEach, // shorthand beforeEach
+  beforeEach,
+  e: expect, // shorthand expect
+  expect,
+  fail,
+  it,
+  fit,
+  xit,
+  pending,
+  spyOn
+};
