+++ tsconfig.json
@@ -17,10 +17,11 @@
     ]
   },
   "include": [
-    "src/**/**.ts"
+    "src/**/*.ts"
   ],
   "exclude": [
-    "src/**/**.spec.ts"
+    "src/**/*.spec.ts",
+    "src/**/*.e2e-spec.ts"
   ],
   "angularCompilerOptions": {
     "genDir": ".",
